+++ PhysIO/code/tapas_physio_fill_empty_parameters.m
@@ -60,7 +60,9 @@ if isempty(physio.log_files.sampling_interval)
             end
         case 'ge'
             physio.log_files.sampling_interval = 25e-3;
-        case {'biopac_mat', 'siemens', 'siemens_tics', 'brainproducts'} % will be read from file later
+        case {'siemens_tics', 'siemens_hcp'}
+             physio.log_files.sampling_interval = 2.5e-3;
+        case {'biopac_mat', 'siemens', 'brainproducts'} % will be read from file later
             physio.log_files.sampling_interval = [];
         otherwise % e.g. custom
             error('Please specify sampling interval for custom text data');