+++ PhysIO/code/tapas_physio_read_physlogfiles_siemens_tics.m
@@ -1,5 +1,5 @@
-function [c, r, t, cpulse, verbose] = tapas_physio_read_physlogfiles_siemens_tics(...
-    log_files, verbose)
+function [c, r, t, cpulse, acq_codes, verbose] = tapas_physio_read_physlogfiles_siemens_tics(...
+    log_files, cardiac_modality, verbose, varargin)
 % reads out physiological time series of Siemens logfiles with Tics
 % The latest implementation of physiological logging in Siemens uses Tics,
 % i.e. time stamps of 2.5 ms duration that are reset every day at midnight.
@@ -7,10 +7,8 @@ function [c, r, t, cpulse, verbose] = tapas_physio_read_physlogfiles_siemens_tic
 % even though individual sampling times may vary - including cardiac,
 % respiratory, pulse oximetry and acquisition time data itself
 %
-%   [cpulse, rpulse, t, c] = tapas_physio_read_physlogfiles_GE(logfiles, ...
-%                               verbose)
-%
-%   NOTE: if one
+% [c, r, t, cpulse, acq_codes, verbose] = tapas_physio_read_physlogfiles_siemens_tics(...
+%    log_files, cardiac_modality, verbose)
 %
 % IN    log_files
 %       .log_cardiac        contains ECG or pulse oximeter time course
@@ -22,18 +20,29 @@ function [c, r, t, cpulse, verbose] = tapas_physio_read_physlogfiles_siemens_tic
 %                           2 entries: 1st entry sampling interval (seconds)
 %                           for cardiac logfile, 2nd entry for respiratory
 %                           logfile
+%                           default: 2.5 ms (1/400 Hz)
+%       cardiac_modality    'ECG' or 'PULS'/'PPU'/'OXY' to determine
+%                           which channel data to be returned
+%                           if not given, will be read out from file name
+%                           suffix
 %       verbose
 %       .level              debugging plots are created if level >=3
 %       .fig_handles        appended by handle to output figure
 %
+%       varargin            propertyName/value pairs, as follows
+%           'ecgChannel'    'v1', 'v2', 'v3', 'v4', 'mean' (default)
+%                           determines which ECG channel to use as
+%                           output cardiac curve
 % OUT
 %   cpulse              time events of R-wave peak in cardiac time series (seconds)
-%                       for GE: usually empty
 %   r                   respiratory time series
 %   t                   vector of time points (in seconds)
-%                       NOTE: This assumes the default sampling rate of 400
-%                       Hz
 %   c                   cardiac time series (ECG or pulse oximetry)
+%   acq_codes           slice/volume start events marked by number <> 0
+%                       for time points in t
+%                       10/20 = scan start/end;
+%                       1 = ECG pulse; 2 = OXY max; 4 = Resp trigger;
+%                       8 = scan volume trigger
 %
 % EXAMPLE
 %   [ons_secs.cpulse, ons_secs.rpulse, ons_secs.t, ons_secs.c] =
@@ -55,6 +64,19 @@ function [c, r, t, cpulse, verbose] = tapas_physio_read_physlogfiles_siemens_tic
 %% read out values
 DEBUG = verbose.level >= 3;
 
+% channel selection different in ECG data, multiple lines for multiple
+% channel, labeled in log file
+switch upper(cardiac_modality)
+    case 'ECG'
+        defaults.ecgChannel = 'v1'; % 'mean'; 'v1'; 'v2', 'v3', 'v4'
+    case {'PPU', 'OXY', 'PMU', 'PULS'}
+        defaults.ecgChannel = 'PULS';
+end
+
+args = tapas_physio_propval(varargin, defaults);
+tapas_physio_strip_fields(args);
+
+
 hasRespirationFile = ~isempty(log_files.respiration);
 hasCardiacFile = ~isempty(log_files.cardiac);
 
@@ -65,6 +87,8 @@ dt = log_files.sampling_interval;
 switch numel(dt)
     case 3
         dtTics = dt(3);
+    case 1
+        dtTics = dt;
     otherwise
         dtTics = 2.5e-3;
 end
@@ -72,20 +96,25 @@ end
 dtCardiac = dtTics;
 dtRespiration = dtTics;
 
+acq_codes = [];
 
 if hasRespirationFile
-    fid = fopen(log_files.respiration);
-    C = textscan(fid, '%d %d %d', 'HeaderLines', 1);
     
-    % check whether textscan worked, otherwise try different format with 4
-    % columns
-    if ~isempty(C{2})
-        r           = double(C{2});
-        rSignals    = double(C{3});
-    else
-        C           = textscan(fid, '%d %s %d %s', 'HeaderLines', 8);
-        r           = double(C{3});
-        rSignals    = ~cellfun(@isempty, C{4});
+    C = tapas_physio_read_files_siemens_tics(log_files.respiration, 'RESP');
+    nColumns = numel(C); % different file formats indicated by diff number of columns
+    
+    extTriggerSignals = [];
+    switch nColumns
+        case 3 % Cologne format
+            r           = double(C{2});
+            rSignals    = double(C{3});
+        case {4,5}
+            r           = double(C{3});
+            rSignals    = ~cellfun(@isempty, C{4});
+    end
+    
+    if nColumns == 5
+        extTriggerSignals = ~cellfun(@isempty, C{5});
     end
     
     rTics           = double(C{1});
@@ -94,43 +123,76 @@ if hasRespirationFile
     
     rpulse          = find(rSignals);
     
+    nSamples        = numel(C{1});
+    racq_codes       = zeros(nSamples,1);
+    
     if ~isempty(rpulse)
+        racq_codes(rpulse) = racq_codes(rpulse) + 4;
         rpulse = tRespiration(rpulse);
     end
+    
+    acqpulse          = find(extTriggerSignals);
+    
+    if ~isempty(acqpulse)
+        racq_codes(acqpulse) = racq_codes(acqpulse) + 8;
+    end
+    
+    
 else
     rpulse          = [];
     r               = [];
     tRespiration    = [];
+    racq_codes      = [];
 end
 
 if hasCardiacFile
-    fid = fopen(log_files.cardiac);
-    C = textscan(fid, '%d %d %d', 'HeaderLines', 1);
     
-    % check whether textscan worked, otherwise try different format with 4
-    % columns
-    if ~isempty(C{2})
-        c           = double(C{2});
-        cSignals    = double(C{3});
-    else
-        C           = textscan(fid, '%d %s %d %s', 'HeaderLines', 8);
-        c           = double(C{3});
-        cSignals    = ~cellfun(@isempty, C{4});
+    C = tapas_physio_read_files_siemens_tics(log_files.cardiac);
+    
+    nColumns = numel(C);
+    
+    % different file formats indicated by diff number of columns
+    
+    extTriggerSignals = [];
+    switch nColumns
+        case 3 % Cologne format
+            c           = double(C{2});
+            cSignals    = double(C{3});
+            cTics           = double(C{1});
+            
+        case {4,5}
+            [cTics, c, cSignals, extTriggerSignals, stringChannels, verbose] = ...
+                tapas_physio_split_data_per_channel_siemens_tics(C, ecgChannel, verbose);
     end
-    cTics           = double(C{1});
+    
+    
     tCardiac        = cTics*dtCardiac ...
         - log_files.relative_start_acquisition;
+    
+    nSamples         = numel(c);
+    cacq_codes       = zeros(nSamples,1);
     
     cpulse          = find(cSignals);
     
     if ~isempty(cpulse)
+        isOxy = any(strfind(upper(log_files.cardiac), 'PULS')); % different codes for PPU
+        
+        cacq_codes(cpulse) = cacq_codes(cpulse) + 1 + isOxy; %+1 for ECG, +2 for PULS
         cpulse = tCardiac(cpulse);
     end
+    
+    acqpulse          = find(extTriggerSignals);
+    
+    if ~isempty(acqpulse)
+        cacq_codes(acqpulse) = cacq_codes(acqpulse) + 8;
+    end
+    
     
 else
     c               = [];
     tCardiac        = [];
     cpulse          = [];
+    cacq_codes      = [];
 end
 
 
@@ -138,8 +200,9 @@ end
 %% interpolate to greater precision, if 2 different sampling rates are given
 
 if DEBUG
-    fh = plot_raw_physlogs(tCardiac, c, tRespiration, r, ...
-        hasCardiacFile, hasRespirationFile, cpulse, rpulse);
+    fh = tapas_physio_plot_raw_physdata_siemens_tics(tCardiac, c, tRespiration, r, ...
+        hasCardiacFile, hasRespirationFile, cpulse, rpulse, cacq_codes, ...
+        racq_codes);
     verbose.fig_handles(end+1) = fh;
 end
 
@@ -147,6 +210,7 @@ end
 hasDifferentSampling = ~isequal(tCardiac, tRespiration);
 
 if hasDifferentSampling && hasCardiacFile && hasRespirationFile
+    %TODO: interpolate acq_codes
     
     nSamplesRespiration = size(r,1);
     nSamplesCardiac = size(c,1);
@@ -157,6 +221,8 @@ if hasDifferentSampling && hasCardiacFile && hasRespirationFile
     if isHigherSamplingCardiac
         t = tCardiac;
         rInterp = interp1(tRespiration, r, t);
+        racq_codesInterp = interp1(tRespiration, racq_codes, t, 'nearest');
+        acq_codes = cacq_codes + racq_codesInterp;
         
         if DEBUG
             fh = plot_interpolation(tRespiration, r, t, rInterp, ...
@@ -168,6 +234,8 @@ if hasDifferentSampling && hasCardiacFile && hasRespirationFile
     else
         t = tRespiration;
         cInterp = interp1(tCardiac, c, t);
+        cacq_codesInterp = interp1(tCardiac, cacq_codes, t, 'nearest');
+        acq_codes = racq_codes + cacq_codesInterp;
         
         if DEBUG
             fh = plot_interpolation(tCardiac, c, t, cInterp, ...
@@ -179,6 +247,18 @@ if hasDifferentSampling && hasCardiacFile && hasRespirationFile
     end
     
 else
+    
+    % merge acq codes
+    if hasCardiacFile
+        if hasRespirationFile
+            acq_codes = cacq_codes + racq_codes;
+        else
+            acq_codes = cacq_codes;
+        end
+    elseif hasRespirationFile
+        acq_codes = racq_codes;
+    end
+    
     nSamples = max(size(c,1), size(r,1));
     t = -log_files.relative_start_acquisition + ((0:(nSamples-1))*...
         min(dtCardiac, dtRespiration))';
@@ -186,47 +266,12 @@ end
 
 end
 
-% Local function to plot raw read-in data;
-function fh = plot_raw_physlogs(tCardiac, c, tRespiration, r, ...
-    hasCardiacFile, hasRespirationFile, cpulse, rpulse)
-fh = tapas_physio_get_default_fig_params();
-stringTitle = 'Siemens Tics - Read-in cardiac and respiratory logfiles';
-set(gcf, 'Name', stringTitle);
-stringLegend = {};
-tOffset = min([tRespiration; tCardiac]);
-if hasCardiacFile
-    plot(tCardiac-tOffset, c, 'r.-'); hold all;
-    stringLegend{1, end+1} =  ...
-        sprintf('Cardiac time course, start time %5.2e', tOffset);
-    
-    if ~isempty(cpulse)
-        stem(cpulse-tOffset, max(abs(c))*ones(size(cpulse)));
-        stringLegend{1, end+1} = 'Detected hearbeats';
-    end
-    
-end
 
-if hasRespirationFile
-    plot(tRespiration-tOffset, r, 'g.-'); hold all;
-    stringLegend{1, end+1} =  ...
-        sprintf('Respiratory time course, start time %5.2e', tOffset);
-    
-    if ~isempty(rpulse)
-        stem(rpulse-tOffset, max(abs(r))*ones(size(rpulse)));
-        stringLegend{1, end+1} = 'Detected Breath starts';
-    end
-    
-    
-end
-xlabel('t (seconds)');
-legend(stringLegend);
-title(stringTitle);
-end
 
 %% Local function to plot interpolation result
 function fh = plot_interpolation(tOrig, yOrig, tInterp, yInterp, ...
     stringOrigInterp)
-fh = tapas_physio_get_default_fig_params;
+fh = tapas_physio_get_default_fig_params();
 stringTitle = sprintf('Interpolation of %s signal', stringOrigInterp{1});
 set(fh, 'Name', stringTitle);
 plot(tOrig, yOrig, 'go--');  hold all;